util: move result<T> into util
